unsafe.Pointer -> []byteのリファクタリング
ページに書き込めてなさそう（どこかでコピーが作成されている？）なので、selectできない

+++ main.go
@@ -139,7 +139,7 @@ func executeSelect(statement Statement, table *Table) ExecuteResult {
 }
 
 func executeInsert(statement Statement, table *Table) ExecuteResult {
-	if table.numRows >= TABLE_MAX_ROWS {
+	if table.numRows >= uint32(TABLE_MAX_ROWS) {
 		return EXECUTE_TABLE_FULL
 	}
